+++ src/slivarpkg/evaluator.nim
@@ -153,11 +153,14 @@ proc set_sample_attributes(ev:Evaluator, by_name: TableRef[string, ISample]) =
       else:
         ev.ctx.pop()
 
-proc trio_kids(samples: seq[Sample]): seq[Sample] =
+proc trio_kids(samples: seq[Sample], by_name: TableRef[string, ISample]): seq[Sample] =
   ## return all samples that have a mom and dad.
   result = newSeqOfCap[Sample](16)
   for sample in samples:
     if sample.mom == nil or sample.dad == nil: continue
+    if sample.id notin by_name: continue
+    if sample.mom.id notin by_name: continue
+    if sample.dad.id notin by_name: continue
     result.add(sample)
 
 proc make_one_row(ev:Evaluator, grps: seq[seq[Sample]], by_name: TableRef[string, ISample]): seq[seq[ISample]] =
@@ -231,7 +234,7 @@ proc newEvaluator*(samples: seq[Sample], groups: seq[Group], float_expressions: 
 
   result.groups = result.make_igroups(groups, by_name, sample_expressions)
 
-  for kid in samples.trio_kids:
+  for kid in samples.trio_kids(by_name):
       result.trios.add([by_name[kid.id], by_name[kid.dad.id], by_name[kid.mom.id]])
 
   result.INFO = result.ctx.newStrictObject("INFO")